Hashing out requirements
The counter microservice seems simple.  I'll set up the database connections
using the GORP 3rd party library.  There's no need (at this point) to log every single request.  So What I'll do is essesntially use the database
as a hash table, checking if the tag exists and incrementing the count
or inserting a new tag entry with a count of 1.  This is a simple
implementation, so don't worry about sharding counters or
concurrency.